pyproject.toml
[tool.poetry]
name = "lpr_project"
version = "0.1.0"
description = ""
authors = ["Maxim Shemyakin <shemmaximg@gmail.com>"]

[tool.poetry.dependencies]
python = "^3.7"
Django = "^2.1"
Pillow = "^5.4"
django-ckeditor = "^5.6"
libsass = "^0.18.0"
django-sass-processor = "^0.7.2"
django-cities-light = "^3.5"
django-uuslug = "^1.1"

[tool.poetry.dev-dependencies]

[build-system]
requires = ["poetry>=0.12"]
build-backend = "poetry.masonry.api"
